admin-backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic[email]>=2.6.0
pydantic-settings>=2.2.0
passlib[bcrypt]>=1.7.4
bcrypt<4.1
python-jose[cryptography]>=3.3.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
loguru>=0.7.0
redis>=5.0.0
